requirements.txt
openai
anthropic
python-dotenv
parea-ai